+++ src/IntermixBase.hpp
@@ -8,7 +8,6 @@ template<int PORTS>
 struct IntermixBase {
     typedef float (*IntermixMatrix)[PORTS];
 	virtual IntermixMatrix expGetCurrentMatrix() { return NULL; }
-	virtual IntermixMatrix expGetMatrix() { return NULL; }
 	virtual int expGetChannelCount() { return 0; }
 };