+++ src/plugin.cpp
@@ -30,6 +30,7 @@ void init(rack::Plugin* p) {
 	p->addModel(modelHive);
 	p->addModel(modelIntermix);
 	p->addModel(modelIntermixGate);
+	p->addModel(modelIntermixEnv);
 	p->addModel(modelSail);
 	p->addModel(modelPile);
 	p->addModel(modelPilePoly);