+++ src/plugin.hpp
@@ -34,6 +34,7 @@ extern Model* modelMaze;
 extern Model* modelHive;
 extern Model* modelIntermix;
 extern Model* modelIntermixGate;
+extern Model* modelIntermixEnv;
 extern Model* modelSail;
 extern Model* modelPile;
 extern Model* modelPilePoly;